refactor(vcs): вынести _RetryTransport и маркер в общий reviewer/vcs/_http
